JAVA-2563: Move OpCode enum to top-level
